T8/requirements.txt
Django==1.10.6
Pillow==4.0.0
django-registration-redux==1.6
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.3
mongoengine==0.11.0
olefile==0.44
pymongo==3.4.0
six==1.10.0
